Delete Sublass and move methods in base class Hopefully fixes problem for CI Server...
