+++ src/test/java/net/sf/jabref/logic/util/io/FileBasedTestCase.java
@@ -1,27 +1,40 @@
-package net.sf.jabref;
+package net.sf.jabref.logic.util.io;
 
+import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.nio.file.Paths;
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
 import java.util.stream.Stream;
 
+import net.sf.jabref.BibtexTestData;
+import net.sf.jabref.Globals;
 import net.sf.jabref.model.database.BibDatabase;
 import net.sf.jabref.model.entry.BibEntry;
 import net.sf.jabref.preferences.JabRefPreferences;
+import net.sf.jabref.support.DevEnvironment;
 
 import org.junit.After;
 import org.junit.Before;
-import static org.junit.Assert.assertTrue;
+import org.junit.Ignore;
+import org.junit.Test;
 
-/**
- * A base class for Testing in JabRef that comes along with some useful
- * functions.
- */
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
+import static org.junit.Assume.assumeFalse;
+
+
 public class FileBasedTestCase {
 
-    protected BibDatabase database;
-    protected BibEntry entry;
-    protected Path rootDir;
+    private BibDatabase database;
+    private BibEntry entry;
+    private Path rootDir;
 
     private String oldPdfDirectory;
     private boolean oldUseRegExp;
@@ -72,6 +85,40 @@ public class FileBasedTestCase {
 
     }
 
+    @Test
+    public void testFindAssociatedFiles() {
+
+        List<BibEntry> entries = Collections.singletonList(entry);
+        List<String> extensions = Arrays.asList("jpg", "pdf");
+        List<File> dirs = Arrays.asList(rootDir.resolve("graphicsDir").toFile(),
+                rootDir.resolve("pdfs").toFile());
+
+        Map<BibEntry, List<File>> results = FileUtil.findAssociatedFiles(entries, extensions, dirs, Globals.prefs);
+
+        assertEquals(2, results.get(entry).size());
+        assertTrue(results.get(entry)
+                .contains(rootDir.resolve(Paths.get("graphicsDir", "subDir", "HipKro03test.jpg")).toFile()));
+        assertFalse(results.get(entry)
+                .contains(rootDir.resolve(Paths.get("graphicsDir", "subDir", "HipKro03test.png")).toFile()));
+        assertTrue(results.get(entry).contains(rootDir.resolve(Paths.get("pdfs", "sub", "HipKro03-sub.pdf")).toFile()));
+    }
+
+    @Test
+    public void testFindFilesException() {
+        List<String> extensions = Arrays.asList("jpg", "pdf");
+        List<File> dirs = Arrays.asList(rootDir.resolve("asdfasdf/asdfasdf").toFile());
+        Set<File> results = FileFinder.findFiles(extensions, dirs);
+
+        assertEquals(0, results.size());
+    }
+
+    @Ignore("Fails on CI Server")
+    @Test(expected = NullPointerException.class)
+    public void testFindFilesNullPointerException() {
+
+        assumeFalse(DevEnvironment.isCIServer());
+        FileFinder.findFiles(null, null);
+    }
     @After
     public void tearDown() {
